modified ZGILS Forcing and Initialization to handle values of CO2 other than increases by powers of 2

+++ Forcing.pxd
@@ -158,7 +158,7 @@ cdef class ForcingZGILS:
     cdef:
         Py_ssize_t loc
         bint varsub
-        Py_ssize_t n_double_co2
+        double n_double_co2
         str reference_type
         double [:] dtdt
         double [:] dqtdt
